FastAPI/requirements.txt
fastapi
uvicorn
python-dotenv
google_streetview
opencv-python